+++ lib/sandal/enc/alg.rb
@@ -6,6 +6,5 @@ module Sandal
   end
 end
 
-require 'sandal/enc/alg/direct'
-require 'sandal/enc/alg/rsa1_5'
-require 'sandal/enc/alg/rsa_oaep'
+require "sandal/enc/alg/direct"
+require "sandal/enc/alg/rsa"