+++ lib/sandal/enc/alg/direct.rb
@@ -1,46 +1,48 @@
-require 'openssl'
+require "openssl"
 
 module Sandal
   module Enc
     module Alg
 
-      # The direct ("dir") key encryption mechanism, which uses a pre-shared 
-      # content master key.
+      # The direct ("dir") key encryption algorithm, which uses a pre-shared symmetric key.
       class Direct
 
-        # @return [String] The JWA name of the algorithm.
-        attr_reader :name
+        # The JWA name of the algorithm.
+        NAME = "dir"
 
-        # @return [String] The pre-shared content master key key.
-        attr_reader :cmk
+        # @return [String] The pre-shared symmetric key.
+        attr_reader :preshared_key
 
-        # Creates a new instance.
+        # Initialises a new instance.
         #
-        # @param cmk [String] The pre-shared content master key.
-        def initialize(cmk)
-          @name = 'dir'
-          @cmk = cmk
+        # @param preshared_key [String] The pre-shared symmetric key.
+        def initialize(preshared_key)
+          @preshared_key = preshared_key
         end
 
-        # Returns an empty string as the content master key is not included in 
-        # the JWE token.
+        # The JWA name of the algorithm.
+        def name
+          NAME
+        end
+
+        # Returns an empty string as the key is not included in JWE tokens using direct key exchange.
         #
-        # @param cmk [String] This parameter is ignored.
+        # @param key [String] This parameter is ignored.
         # @return [String] An empty string.
-        def encrypt_cmk(cmk)
-          ''
+        def encrypt_key(key)
+          ""
         end
 
-        # Returns the pre-shared content master key.
+        # Returns the pre-shared content key.
         #
-        # @param encrypted_cmk [String] The encrypted content master key.
-        # @return [String] The pre-shared content master key.
-        # @raise [Sandal::TokenError] encrypted_cmk is not nil or empty.
-        def decrypt_cmk(encrypted_cmk)
-          unless encrypted_cmk.nil? || encrypted_cmk.empty?
-            raise Sandal::TokenError, 'Token must not include encrypted CMK.' 
+        # @param encrypted_key [String] The encrypted key.
+        # @return [String] The pre-shared symmetric key.
+        # @raise [Sandal::InvalidTokenError] encrypted_key is not nil or empty.
+        def decrypt_key(encrypted_key)
+          unless encrypted_key.nil? || encrypted_key.empty?
+            raise Sandal::InvalidTokenError, "Tokens using direct key exchange must not include a content key."
           end
-          @cmk
+          @preshared_key
         end
 
       end