+++ spec/sandal/enc/alg/direct_spec.rb
@@ -1,53 +1,47 @@
-require 'helper'
-require 'openssl'
+require "helper"
+require "openssl"
 
 describe Sandal::Enc::Alg::Direct do
 
-  context '#name' do
-
-    it 'is "dir"' do
-      alg = Sandal::Enc::Alg::Direct.new('some cmk')
-      alg.name.should == 'dir'
+  context "#name" do
+    it "is 'dir'" do
+      alg = Sandal::Enc::Alg::Direct.new("some key")
+      alg.name.should == "dir"
     end
-
   end
 
-  context '#cmk' do
-
-    it 'returns the real CMK' do
-      cmk = 'the real cmk'
-      alg = Sandal::Enc::Alg::Direct.new(cmk)
-      alg.cmk.should == cmk
+  context "#preshared_key" do
+    it "returns the pre-shared key" do
+      key = "the pre-shared key"
+      alg = Sandal::Enc::Alg::Direct.new(key)
+      alg.preshared_key.should == key
     end
-
   end
 
-  context '#encrypt_cmk' do
-
-    it 'returns an empty string' do
-      alg = Sandal::Enc::Alg::Direct.new('the real cmk')
-      alg.encrypt_cmk('any value').should == ''
+  context "#encrypt_key" do
+    it "returns an empty string" do
+      alg = Sandal::Enc::Alg::Direct.new("the real key")
+      alg.encrypt_key("any value").should == ""
     end
-
   end
 
-  context '#decrypt_cmk' do
+  context "#decrypt_key" do
 
-    it 'returns the real CMK when the value to decrypt is nil' do
-      cmk = 'the real cmk'
-      alg = Sandal::Enc::Alg::Direct.new(cmk)
-      alg.decrypt_cmk(nil).should == cmk
+    it "returns the pre-shared content key when the value to decrypt is nil" do
+      key = "a pre-shared key"
+      alg = Sandal::Enc::Alg::Direct.new(key)
+      alg.decrypt_key(nil).should == key
     end
 
-    it 'returns the real CMK when the value to decrypt is empty' do
-      cmk = 'the real cmk'
-      alg = Sandal::Enc::Alg::Direct.new(cmk)
-      alg.decrypt_cmk('').should == cmk
+    it "returns the pre-shared content key when the value to decrypt is empty" do
+      key = "my pre-shared key"
+      alg = Sandal::Enc::Alg::Direct.new(key)
+      alg.decrypt_key("").should == key
     end
 
-    it 'raises a TokenError if the value to decrypt is not nil or empty' do
-      alg = Sandal::Enc::Alg::Direct.new('the real cmk')
-      expect { alg.decrypt_cmk('a value') }.to raise_error Sandal::TokenError
+    it "raises an InvalidTokenError if the value to decrypt is not nil or empty" do
+      alg = Sandal::Enc::Alg::Direct.new("the pre-shared key")
+      expect { alg.decrypt_key("a value") }.to raise_error Sandal::InvalidTokenError
     end
 
   end